+++ packages/components/src/templates/next/layouts/Collection/Collection.tsx
@@ -12,7 +12,8 @@ import {
 } from "~/utils"
 import { Skeleton } from "../Skeleton"
 import CollectionClient from "./CollectionClient"
-import { getAvailableFilters, shouldShowDate } from "./utils"
+import { getAvailableFilters } from "./filterUtils"
+import { shouldShowDate } from "./utils"
 
 const CATEGORY_OTHERS = "Others"
