+++ packages/components/src/templates/next/layouts/Collection/filterUtils/index.ts
@@ -1,4 +1,2 @@
-export * from "./getCategoryFilter"
-export * from "./getYearFilter"
-export * from "./getTagFilters"
 export * from "./constants"
+export * from "./getAvailableFilters"
